+++ src/main/java/com/argusoft/form/repository/FillFormRepository.java
@@ -20,13 +20,17 @@ public interface FillFormRepository extends JpaRepository<FillForm, Long> {
         @Modifying
         @Query(value = "INSERT INTO fill_form (form_id, user_id, answers, created_at, location, location_id) VALUES (:formId, :userId, CAST(:answers AS jsonb), :createdAt, CAST(:location AS loc), :locationId)", nativeQuery = true)
         void insertFillForm(Long formId, Long userId, String answers, LocalDateTime createdAt, String location,
-                        Integer locationId);
+                        Long locationId);
 
         // @Transactional
         // @Modifying
-        // @Query(value = "UPDATE fill_form SET form_id = :formId, user_id = :userId, answers = CAST(:answers AS jsonb), created_at = :createdAt, location = :location, location_id = :locationId  WHERE form_id = :formId AND user_id = :userId", nativeQuery = true)
-        // void updateFillForm(Long formId, Long userId, String answers, LocalDateTime createdAt, String location,
-        //                 Integer locationId);
+        // @Query(value = "UPDATE fill_form SET form_id = :formId, user_id = :userId,
+        // answers = CAST(:answers AS jsonb), created_at = :createdAt, location =
+        // :location, location_id = :locationId WHERE form_id = :formId AND user_id =
+        // :userId", nativeQuery = true)
+        // void updateFillForm(Long formId, Long userId, String answers, LocalDateTime
+        // createdAt, String location,
+        // Integer locationId);
 
         @Query("SELECT f FROM FillForm f")
         Page<FillForm> getAllFillForm(Pageable p);
@@ -36,7 +40,7 @@ public interface FillFormRepository extends JpaRepository<FillForm, Long> {
 
         @Transactional
         @Modifying
-        @Query(value = "DELETE FROM fill_form WHERE id = :id", nativeQuery=true)
+        @Query(value = "DELETE FROM fill_form WHERE id = :id", nativeQuery = true)
         void deleteFillFormById(Long id);
 
 }
